Create importRedcap.R
Joann Alvarez did the heavy lifting on this one.  I based this on one of her functions to parse the metadata and write the function calls to factor() and label().  I don't know if you think this fits in your package, but I wanted to share in case you thought it was useful.