Factor levels for 'checkbox' type
Since the d argument in fieldToVars takes a vector for the .redcapApiConnection method, the variables that are part of a 'checkbox' type (check all that apply) can't be processed as directly as other types.  Lines 241-250 expand the variable names to include all of the options for the checkbox variables (with the ___x suffix).  Line 264 had to be amended to find the right meta_data for those variables (without the ___x suffix).  This will work for as many as 1,000,000,000 options.

I made no changes to the .redcapDbConnection method and so results may vary.