Allow character encodings
The usual practice in REDCap is to encode categorical variables as
1, Option A
2, Option B

However, the encodings
a, Option A
b, Option B

are also allowed.  Underscores are allowed, as well, but the changes on this commit do not accomodate the underscores.  I'll probably come back to that after it bites me in the butt.