Update syncUnderscoreCodings.R
made an adjustment for the import.  While the underscores are stripped out on export, the import still requires the underscores in the variable names.  For the import, the meta_data is not changed, and then the 'checkbox_field_name_map' can be used to correct field names in the data to match the meta_data.